Upgrade RQ to v4
