+++ kafka-ui-react-app/package.json
@@ -12,6 +12,7 @@
     "@hookform/resolvers": "^2.7.1",
     "@reduxjs/toolkit": "^1.8.3",
     "@rooks/use-outside-click-ref": "^4.10.1",
+    "@tanstack/react-query": "^4.0.5",
     "@testing-library/react": "^13.2.0",
     "@types/testing-library__jest-dom": "^5.14.5",
     "@types/yup": "^0.29.13",
@@ -35,7 +36,6 @@
     "react-hook-form": "7.6.9",
     "react-is": "^18.2.0",
     "react-multi-select-component": "^4.0.6",
-    "react-query": "^3.39.1",
     "react-redux": "^8.0.2",
     "react-router-dom": "^6.3.0",
     "redux": "^4.2.0",
@@ -106,7 +106,6 @@
     "eslint-plugin-prettier": "^4.0.0",
     "eslint-plugin-react": "^7.29.4",
     "eslint-plugin-react-hooks": "^4.5.0",
-    "fetch-mock-jest": "^1.5.1",
     "husky": "^8.0.1",
     "jest-environment-jsdom": "^28.1.1",
     "jest-sonar-reporter": "^2.0.0",
