+++ kafka-ui-react-app/src/lib/fixtures/brokers.ts
@@ -5,21 +5,6 @@ export const brokersPayload = [
   { id: 2, host: 'b-2.test.kafka.amazonaws.com', port: 9092 },
 ];
 
-export const clusterStatsPayload = {
-  brokerCount: 2,
-  activeControllers: 1,
-  onlinePartitionCount: 138,
-  offlinePartitionCount: 0,
-  inSyncReplicasCount: 239,
-  outOfSyncReplicasCount: 0,
-  underReplicatedPartitionCount: 0,
-  diskUsage: [
-    { brokerId: 0, segmentSize: 334567, segmentCount: 245 },
-    { brokerId: 1, segmentSize: 12345678, segmentCount: 121 },
-  ],
-  version: '2.2.1',
-};
-
 const partition = {
   broker: 2,
   offsetLag: 0,
