+++ kafka-ui-react-app/src/lib/hooks/api/__tests__/kafkaConnect.spec.ts
@@ -3,7 +3,7 @@ import { renderQueryHook, TestQueryClientProvider } from 'lib/testHelpers';
 import * as hooks from 'lib/hooks/api/kafkaConnect';
 import fetchMock from 'fetch-mock';
 import { connectors, connects, tasks } from 'lib/fixtures/kafkaConnect';
-import { UseQueryResult } from 'react-query';
+import { UseQueryResult } from '@tanstack/react-query';
 import { ConnectorAction } from 'generated-sources';
 
 const clusterName = 'test-cluster';
