+++ kafka-ui-react-app/src/lib/hooks/api/brokers.ts
@@ -1,5 +1,5 @@
 import { brokersApiClient as api } from 'lib/api';
-import { useQuery } from 'react-query';
+import { useQuery } from '@tanstack/react-query';
 import { ClusterName } from 'redux/interfaces';
 
 export function useBrokers(clusterName: ClusterName) {
